Add notes from today's discussion
Notes from today's discussion with Mike Barlage, Erik Kluzek, Negin
Sobhani, Sean Swenson and Mariana Vertenstein.

In addition to the notes in the README, note the suggested subroutine
renames in the chosen solution, and the fact that we should pass 'col'
into this routine rather than having it accessed via 'use'.