Start work on solution with explicit array args for ALL routines
This adds long argument lists for the tracer-only routines, but keeps
the tracer-only routines consistent with the bulk-only or
bulk-and-tracer routines. (I have only implemented this for the first
few routines.)

In addition to the consistency, I like this because I think this will
make it more likely for the sources (for state updates & tracer updates)
to stay in sync: I think it will be easier for someone to notice that
they need to change the argument list for the tracer flux update
routine.